Add batch_run_type for running particular job types
For example, this is very important to quickly run `ticker` job types
regularly
